Fix `find_entry_point` when the bounding box is very large
Because we spline the poloidal flux along a parameterised line that
stretches from the launch position to the *other* side of the bounding
box, very large bounding boxes may cause our initial set of points for
this spline to have very few or even no points inside the plasma. In
this case, there will be no roots of the spline.

The solution is to make sure we have enough points that we're
guaranteed to have at least some inside the plasma.